# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Janeiro Tour API - Premium Brazil Travel Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tours:
    get:
      operationId: listTours
      tags: [tours]
      summary: List all tours
      parameters:
        - name: destination
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [private, group, all]
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
    post:
      operationId: createTour
      tags: [tours]
      summary: Create a tour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
...
  /tours/{id}:
    get:
      operationId: getTour
      tags: [tours]
      summary: Get a tour by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tour detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        "404":
          description: Not found
    patch:
      operationId: updateTour
      tags: [tours]
      summary: Update a tour
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
    delete:
      operationId: deleteTour
      tags: [tours]
      summary: Delete a tour
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tours/{id}/slug/{slug}:
    get:
      operationId: getTourBySlug
      tags: [tours]
      summary: Get a tour by slug
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tour detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"

  # Destinations
...
components:
  schemas:
...
    Tour:
      type: object
      required: [id, slug, title, destination, category, durationHours, priceFrom, currency, tourType, languages, groupSizeMax, imageUrl, highlights, overview, includedItems, notIncludedItems, itinerary, meetingPoint, cancellationPolicy, published, featured, sortOrder, createdAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        titleEs:
          type: ["string", "null"]
        titlePt:
          type: ["string", "null"]
        destination:
          type: string
        category:
          type: string
        durationHours:
          type: number
        priceFrom:
          type: number
        currency:
          type: string
        tourType:
          type: string
          enum: [private, group, both]
        languages:
          type: array
          items:
            type: string
        groupSizeMax:
          type: integer
        imageUrl:
          type: string
        galleryImages:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        overview:
          type: string
        overviewEs:
          type: ["string", "null"]
        overviewPt:
          type: ["string", "null"]
        includedItems:
          type: array
          items:
            type: string
        notIncludedItems:
          type: array
          items:
            type: string
        itinerary:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryStep"
        meetingPoint:
          type: string
        cancellationPolicy:
          type: string
        regiondoWidget:
          type: ["string", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        published:
          type: boolean
        featured:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
    TourInput:
      type: object
      required: [slug, title, destination, category, durationHours, priceFrom, currency, tourType, languages, groupSizeMax, imageUrl, highlights, overview, includedItems, notIncludedItems, itinerary, meetingPoint, cancellationPolicy]
      properties:
        slug:
          type: string
        title:
          type: string
        titleEs:
          type: string
        titlePt:
          type: string
        destination:
          type: string
        category:
          type: string
        durationHours:
          type: number
        priceFrom:
          type: number
        currency:
          type: string
        tourType:
          type: string
          enum: [private, group, both]
        languages:
          type: array
          items:
            type: string
        groupSizeMax:
          type: integer
        imageUrl:
          type: string
        galleryImages:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        overview:
          type: string
        overviewEs:
          type: string
        overviewPt:
          type: string
        includedItems:
          type: array
          items:
            type: string
        notIncludedItems:
          type: array
          items:
            type: string
        itinerary:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryStep"
        meetingPoint:
          type: string
        cancellationPolicy:
          type: string
        regiondoWidget:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
        sortOrder:
          type: integer
    TourUpdate:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        titleEs:
          type: string
        titlePt:
          type: string
        destination:
          type: string
        category:
          type: string
        durationHours:
          type: number
        priceFrom:
          type: number
        currency:
          type: string
        tourType:
          type: string
        languages:
          type: array
          items:
            type: string
        groupSizeMax:
          type: integer
        imageUrl:
          type: string
        galleryImages:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        overview:
          type: string
        overviewEs:
          type: string
        overviewPt:
          type: string
        includedItems:
          type: array
          items:
            type: string
        notIncludedItems:
          type: array
          items:
            type: string
        itinerary:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryStep"
        meetingPoint:
          type: string
        cancellationPolicy:
          type: string
        regiondoWidget:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
        sortOrder:
          type: integer
    ItineraryStep:
      type: object
      required: [order, title, description]
      properties:
        order:
          type: integer
        title:
          type: string
        description:
          type: string
        duration:
          type: ["string", "null"]
...