# lib/api-zod/src/generated/types/tourInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Janeiro Tour API - Premium Brazil Travel Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryStep } from './itineraryStep';
import type { TourInputTourType } from './tourInputTourType';

export interface TourInput {
  slug: string;
  title: string;
  titleEs?: string;
  titlePt?: string;
  destination: string;
  category: string;
  durationHours: number;
  priceFrom: number;
  currency: string;
  tourType: TourInputTourType;
  languages: string[];
  groupSizeMax: number;
  imageUrl: string;
  galleryImages?: string[];
  highlights: string[];
  overview: string;
  overviewEs?: string;
  overviewPt?: string;
  includedItems: string[];
  notIncludedItems: string[];
  itinerary: ItineraryStep[];
  meetingPoint: string;
  cancellationPolicy: string;
  regiondoWidget?: string;
  seoTitle?: string;
  seoDescription?: string;
  published?: boolean;
  featured?: boolean;
  sortOrder?: number;
}